requirements.txt
beautifulsoup4==4.4.1
cssmin==0.2.0
Flask==0.10.1
Flask-Assets==0.11
Flask-Script==2.0.5
Frozen-Flask==0.12
hyphenator==0.5.1
itsdangerous==0.24
Jinja2==2.8
Markdown==2.6.6
MarkupSafe==0.23
Pygments==2.1.3
PyHyphen==2.0.5
pyScss==1.3.4
pytils==0.3
six==1.10.0
webassets==0.11.1
Werkzeug==0.11.9
